moved resolve to its own module
