+++ os/contiki-net.h
@@ -46,11 +46,7 @@
 
 #include "net/ipv6/uip-icmp6.h"
 #include "net/ipv6/uip-ds6.h"
-
-#include "net/ipv6/resolv.h"
-
 #include "net/ipv6/psock.h"
-
 #include "net/ipv6/udp-socket.h"
 #include "net/ipv6/tcp-socket.h"
