+++ os/net/app-layer/http-socket/Makefile.http-socket
@@ -1,0 +1,1 @@
+MODULES += os/services/resolv
