+++ os/services/resolv/module-macros.h
@@ -1,0 +1,1 @@
+#define BUILD_WITH_RESOLV 1
